backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
python-dotenv==1.0.1
supabase==2.8.1
google-generativeai==0.8.3
